+++ linkerd/core/src/main/scala/io/buoyant/linkerd/Linker.scala
@@ -100,7 +100,13 @@ object Linker {
 
       val telemeters = telemetry match {
         case None => Seq(defaultTelemeter)
-        case Some(telemeters) => telemeters.map(_.mk(Stack.Params.empty))
+        case Some(telemeters) => telemeters.map {
+          case t if t.disabled =>
+            val msg = s"The ${t.getClass.getCanonicalName} telemeter is experimental and must be " +
+              "explicitly enabled by setting the `experimental' parameter to `true'."
+            throw new IllegalArgumentException(msg) with NoStackTrace
+          case t => t.mk(Stack.Params.empty)
+        }
       }
 
       // Telemeters may provide StatsReceivers.