+++ CHANGES.md
@@ -1,3 +1,7 @@
+## In the next release...
+
+* Add experimental StatsD telemeter
+
 ## 0.8.5 2017-01-06
 
 * Introduce the grpc-gen and grpc-runtime projects, enabling code
@@ -14,7 +18,7 @@
   `dtab-local`
 * Transformers:
   * Transformers will now prepend a prefix to the id of the bound names they
-    modify. 
+    modify.
   * Fix localhost transformer when used on systems with unresolvable hostname.
 
 ## 0.8.4 2016-12-05